TypeScript/resources/code/ts_lib/Apollo-Latest/apollographql/internal-platform-orb/src/scripts/circleci-job-canceller/requirements.txt
requests
python-dateutil
chevron
